Day 07

+++ template/src/main.rs
@@ -1,4 +1,4 @@
-use common::load_data;
+use common::load_data_full;
 
 fn main() {
     // let input: String = load_data_full("data/day_0.txt");
@@ -6,11 +6,13 @@ fn main() {
     // println!("Day  Part 2: {}", part_2(&input));
 }
 
-fn part_1(input: &str) -> u32 {
-    input.len() as u32
+fn part_1(input: &str) -> usize {
+    input.len()
 }
 
-fn part_2(input: &str) -> u32 { input.len() as u32 }
+fn part_2(input: &str) -> usize {
+    input.len()
+}
 
 #[cfg(test)]
 mod tests {
